requirements.txt
httpx<0.28
tqdm>=4.66.5
yt-dlp>=2025.12.8
youtube-search-python>=1.6.6
ffmpeg-python>=0.2.0
